Updated filenames. Worked on PiParser registering
This is a big update. I renamed and tried to reorganize the custom
parser structure.
##Parsers
Currently, all parsers must inherit from CustomParser and implement the
parse method.
###CustomBufferParser inherits from CustomParser and implements the
parse method. This method converts the char* object into a (currently
untested) Google Protocol Buffer message. This is a virtual class and
classes that inherit from it must implement the parse method. This
method gets passed a ProtocolBuffer object and must return one as well.