Made the PiParser a normal class (instead of a singleton)
A PiParser singleton class doesn’t make sense anymore this far into the project. For group registration, the group registration parser will need to be given an instance of a client manager where groups will be registered. If the PiParser was a singleton, we would be unable to have more than one instance of the group registration parser which means having more than one client manager would be pointless. By removing the singleton, we are able to run multiple instances of the default PiServer in the same program without an issue.